backend/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
facenet-pytorch
opencv-python-headless
Pillow
numpy
python-multipart
